Move Exception benchmarks from FileSystem.Handle to other modules
