+++ benchmark/Streamly/Benchmark/FileSystem/Handle.hs
@@ -33,7 +33,10 @@ import qualified Handle.Read as RO
 import Data.IORef
 import Gauge hiding (env)
 import Streamly.Benchmark.Common
-import Handle.Common
+-- TBD Change it to point to Common in parent directory
+-- Done
+-- import Handle.Common
+import Streamly.Benchmark.CommonH
 
 -------------------------------------------------------------------------------
 --
@@ -46,11 +49,11 @@ moduleName = "FileSystem.Handle"
 --
 -------------------------------------------------------------------------------
 
-smallFileSize :: Int
-smallFileSize = 10 * 1024 * 1024
+-- smallFileSize :: Int
+-- smallFileSize = 10 * 1024 * 1024
 
-bigFileSize :: Int
-bigFileSize = 100 * 1024 * 1024
+-- bigFileSize :: Int
+-- bigFileSize = 100 * 1024 * 1024
 
 -------------------------------------------------------------------------------
 --
@@ -59,50 +62,51 @@ bigFileSize = 100 * 1024 * 1024
 main :: IO ()
 main = do
     (_, cfg, benches) <- parseCLIOpts defaultStreamSize
-    r <- lookupEnv "Benchmark_FileSystem_Handle_InputFile"
-    (small, big) <-
-        case r of
-            Just inFileName -> return (inFileName, inFileName)
-            Nothing -> do
-                -- XXX will this work on windows/msys?
-                let cmd infile size =
-                        "mkdir -p " ++ scratchDir
-                            ++ "; test -e " ++ infile
-                            ++ " || { echo \"creating input file " ++ infile
-                            ++ "\" && head -c " ++ show size
-                            ++ " </dev/urandom >" ++ infile
-                            ++ ";}"
-                runProcess_ (shell (cmd inFileSmall smallFileSize))
-                runProcess_ (shell (cmd inFileBig bigFileSize))
-                return (inFileSmall, inFileBig)
+    env <- mkBenchEnv "Benchmark_FileSystem_Handle_InputFile"
+    -- r <- lookupEnv "Benchmark_FileSystem_Handle_InputFile"
+    -- (small, big) <-
+    --     case r of
+    --         Just inFileName -> return (inFileName, inFileName)
+    --         Nothing -> do
+    --             -- XXX will this work on windows/msys?
+    --             let cmd infile size =
+    --                     "mkdir -p " ++ scratchDir
+    --                         ++ "; test -e " ++ infile
+    --                         ++ " || { echo \"creating input file " ++ infile
+    --                         ++ "\" && head -c " ++ show size
+    --                         ++ " </dev/urandom >" ++ infile
+    --                         ++ ";}"
+    --             runProcess_ (shell (cmd inFileSmall smallFileSize))
+    --             runProcess_ (shell (cmd inFileBig bigFileSize))
+    --             return (inFileSmall, inFileBig)
 
-    putStrLn $ "Using small input file: " ++ small
-    smallHandle <- openFile small ReadMode
+    -- putStrLn $ "Using small input file: " ++ small
+    -- smallHandle <- openFile small ReadMode
 
-    putStrLn $ "Using big input file: " ++ big
-    bigHandle <- openFile big ReadMode
+    -- putStrLn $ "Using big input file: " ++ big
+    -- bigHandle <- openFile big ReadMode
 
-    putStrLn $ "Using output file: " ++ outfile
-    outHandle <- openFile outfile WriteMode
-    devNull <- openFile "/dev/null" WriteMode
+    -- putStrLn $ "Using output file: " ++ outfile
+    -- outHandle <- openFile outfile WriteMode
+    -- devNull <- openFile "/dev/null" WriteMode
 
-    ssize <- fromIntegral <$> getFileSize small
-    bsize <- fromIntegral <$> getFileSize big
+    -- ssize <- fromIntegral <$> getFileSize small
+    -- bsize <- fromIntegral <$> getFileSize big
 
-    ref <- newIORef $ RefHandles
-        { smallInH = smallHandle
-        , bigInH = bigHandle
-        , outputH = outHandle
-        }
+    -- ref <- newIORef $ RefHandles
+    --     { smallInH = smallHandle
+    --     , bigInH = bigHandle
+    --     , outputH = outHandle
+    --     }
 
-    let env = BenchEnv
-            { href = ref
-            , smallSize = ssize
-            , bigSize = bsize
-            , nullH = devNull
-            , smallInFile = small
-            , bigInFile = big
-            }
+    -- let env = BenchEnv
+    --         { href = ref
+    --         , smallSize = ssize
+    --         , bigSize = bsize
+    --         , nullH = devNull
+    --         , smallInFile = small
+    --         , bigInFile = big
+    --         }
 
     runMode (mode cfg) cfg benches (allBenchmarks env)
